+++ src/main/java/me/reklessmitch/mitchprisonscore/mitchprofiles/cmds/currency/CmdCurrencyRemove.java
@@ -11,6 +11,8 @@ import me.reklessmitch.mitchprisonscore.mitchprofiles.configs.ProfilePlayer;
 import me.reklessmitch.mitchprisonscore.mitchprofiles.utils.CurrencyUtils;
 import org.bukkit.entity.Player;
 
+import java.math.BigInteger;
+
 public class CmdCurrencyRemove extends CurrencyCommands {
 
     public CmdCurrencyRemove() {
@@ -27,7 +29,7 @@ public class CmdCurrencyRemove extends CurrencyCommands {
         Player player = this.readArg();
         String currency = this.readArg();
         String amount = this.readArg();
-        long amountInt = CurrencyUtils.parse(amount);
+        BigInteger amountInt = CurrencyUtils.parse(amount);
         ProfilePlayer pp = ProfilePlayer.get(player.getUniqueId());
         pp.getCurrency(currency).take(amountInt);
         pp.changed();
